move paper options rendering to util, organize preload for clearer reading
