+++ src/preload.js
@@ -2,36 +2,26 @@ import { Book, pagesizes } from './book.js';
 import { loadForm } from './utils/formUtils.js';
 import { handleFileChange, handleInputChange } from './utils/changeHandlers.js';
 import { handleGenerateClick } from './utils/clickHandlers.js';
+import { renderPaperSelectOptions } from './utils/renderUtils.js';
 
 window.addEventListener('DOMContentLoaded', () => {
-    const paperlist = document.getElementById('paper_size');
     const generate = document.getElementById('generate');
+    const bookbinderForm = document.getElementById('bookbinder');
+    const fileInput = document.getElementById('input_file');
+    const inputs = document.querySelectorAll('input, select');
 
-    let bookbinderForm = document.getElementById('bookbinder');
-
-    Object.keys(pagesizes).forEach((key) => {
-        let opt = document.createElement('option');
-        opt.setAttribute('value', key);
-        if (key == 'A4') {
-            opt.setAttribute('selected', true);
-        }
-        opt.innerText = key;
-        paperlist.appendChild(opt);
-    });
-
+    renderPaperSelectOptions();
     loadForm();
 
-    let book = new Book();
-    let inputs = document.querySelectorAll('input, select');
+    const book = new Book();
+    
     inputs.forEach((input) => {
         input.addEventListener('change', () =>
             handleInputChange(book, bookbinderForm)
         );
     });
 
-    document
-        .querySelector('#input_file')
-        .addEventListener('change', (e) => handleFileChange(e, book));
+    fileInput.addEventListener('change', (e) => handleFileChange(e, book));
 
     generate.addEventListener('click', () =>
         handleGenerateClick(generate, book)
