+++ src/utils/renderUtils.js
@@ -1,3 +1,5 @@
+import { pagesizes } from '../book';
+
 export function renderPageCount(book) {
     document.getElementById('page_count').innerText = book.pagecount;
 }
@@ -10,11 +12,24 @@ export function renderInfoBox(book) {
 
     const outputPages = book.book.pagelist.reduce((acc, list) => {
         list.forEach((sublist) => (acc += sublist.length));
-		return acc;
+        return acc;
     }, 0);
 
     totalSheets.innerText = book.book.sheets;
     sigCount.innerText = book.book.sigconfig.length;
     sigArrange.innerText = book.book.sigconfig.join(', ');
     totalPages.innerText = outputPages;
+}
+
+export function renderPaperSelectOptions() {
+    const paperList = document.getElementById('paper_size');
+    Object.keys(pagesizes).forEach((key) => {
+        let opt = document.createElement('option');
+        opt.setAttribute('value', key);
+        if (key == 'A4') {
+            opt.setAttribute('selected', true);
+        }
+        opt.innerText = key;
+        paperList.appendChild(opt);
+    });
 }
